Add run_problem.sh script
Move the run_problem.sh from our old propeller repo into this repo in order to clean up the old repo and transition into Lee Spector's version of propeller. This method takes six arguments:

The first argument is the path of the problem. An example of this is propeller.problems.PSB2.middle-character
The second argument is the path of the folder you would like to put run files into
The third argument is the generic name for all text files. For example, middle_char will turn into middle_charX.txt, where X is an integer
The fourth argument is the starting number to start the counting of X
The fifth argument is the number of runs to perform at once. This number should be limited depending on the amount of power needed for that specific problem.
The final argument is the number of iterations youd like to run.

Full example: ./run_problem.sh propeller.software.PSB2.middle-aharacter middle_char_folder middle_char_text 0 2 4 will generate four iterations of 2 runs each for the middle-character problem.

./run_problem.sh