Modify run_problem.sh script
Changed the run_problem.sh script to pair with the modification to propeller on 3/5/2022 that added simplification to the end of each run along with outputting the correct-output, behavior, error, and steps-remaining.

This adds two new parameters to the scripts:

The 7th argument now determines the amount of k-steps the user wants while simplifying.
The 8th argument determines the amount of simplification steps; which is the number of steps you want to reduce to if possible.