Add new results and functions
Added some new functions that automate the process of creating graphs on a given dataset. The main function, GenerateGraphs, takes a dataset (givenData), the initial step-limit (originalStepLimit), and the name of the file you'd like to output to (saveToFile). It also has an optional argument (groupPlots) of whether to export all of the graphs into the same file (TRUE, default) or not (FALSE).

This function first wrangles all of the data into the proper classes (from the WrangleData function). Then, it prints a graph of all runs that generalized at the given step limit argument (from the GraphOriginalGeneralizations function), a graph of all runs that generalize anywhere not at the step-limit argument (from the GraphOtherGeneralizations function), and a graph which shows all of the runs still running out of steps by the end of the step-limit range (from the GraphRunningOutOfSteps function) and are below a specific threshold to determine whether they were close to solving the problem (ideally 5% of the total number of test cases).

Additionally, this commit fixes multiple typos in original datasets such as having duplicated headers and therefore giving NA values. Lastly, this commit adds in two new datasets: XOResultsMC.txt and MutLIOZResults.txt.

XOResultsMC.txt contains 30 runs of the middle-character problem where crossover (XO) is included as a possibility (0.5 umad and 0.5 XO).

MutLIOZResults.txt contains 30 runs of the last-index-of-zero problem where only umad is an option (1.0 umad and 0.0 XO).

An example of how to use the GenerateGraphs function would be:

GenerateGraphs(MiddleCharData, 2000, "middleChar.pdf")

This would output the graphs described above for the dataset "MiddleCharData", which was originally run on a step-limit of 2000. Additionally, it will output the graphs to the file "middleChar.pdf". Since the optional argument wasn't passed, it defaults to TRUE and all graphs will be exported to the same file.