Experiment with using paramters; rename FailedCases
This renames "dataset$FailedCases" to
"dataset$NumFailedCases" to improve readability.

This also experiments with using heading parameters
to specify the filename and step limit for a
dataset. The goal would be to create a new Markdown
document that is parameterized like this and
generates a complete report on that dataset so we
can get consistent, automated reporting on all our
datasets.