Create a "generic" step limit Rmarkdown file
Richard Lussier and I wrote a start on a "generic"
RMarkdown document that we can use to generate the
basic graphs and counting statistics for our
datasets. The idea is that we can use the `render`
function in the RMarkdown package to run this
over and over on multiple datasets to get a
collection of standard output files for each run.

Co-authored-by: RichardLussier <lussi036@morris.umn.edu>