Add new dataset and graphs
Imported in the new "MCnoCombRes.txt" dataset, which runs the "middle character" problem without the use of the y-combinator instruction. This commit also generates the separate graphs for this dataset, although only one is included since the other two were empty.

Additionally, this commit deletes a graph, "3_MutLIOZgraphs.png" because, similarly to the previous explanation, it is empty and therefore is not needed.