Add facet plot code
This commit adds a draft for the facet plot code. This code is mostly complete (for LIOZ), but has some issues. The main issue that still needs to be fixed is that the insets need to be independent from each other (each inset should zoom on different parts of their respective plot).